move tf utils
